:recycle: api: Move attendance sheet to enrolment
