+++ api/tests/certification/enrolment/unit/domain/usecases/delete-session_test.js
@@ -1,4 +1,4 @@
-import { deleteSession } from '../../../../../../src/certification/enrollment/domain/usecases/delete-session.js';
+import { deleteSession } from '../../../../../../src/certification/enrolment/domain/usecases/delete-session.js';
 import { SessionStartedDeletionError } from '../../../../../../src/certification/session/domain/errors.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
