+++ api/tests/unit/domain/usecases/flag-session-results-as-sent-to-prescriber_test.js
@@ -1,6 +1,6 @@
 import { NotFoundError } from '../../../../lib/domain/errors.js';
 import { flagSessionResultsAsSentToPrescriber } from '../../../../lib/domain/usecases/flag-session-results-as-sent-to-prescriber.js';
-import { Session } from '../../../../src/certification/enrollment/domain/models/Session.js';
+import { Session } from '../../../../src/certification/enrolment/domain/models/Session.js';
 import { catchErr, expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | UseCase | flag-session-results-as-sent-to-prescriber', function () {
